backend/requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
google-genai
